Added road

+++ Raiden Project/ModuleRender.cpp
@@ -79,7 +79,7 @@ bool ModuleRender::CleanUp()
 }
 
 // Blit to screen
-bool ModuleRender::Blit(SDL_Texture* texture, int x, int y, SDL_Rect* section, int w, int h)
+bool ModuleRender::Blit(SDL_Texture* texture, float x, float y, SDL_Rect* section, int w, int h)
 {
 	bool ret = true;
 	SDL_Rect rect;
